+++ medusa/indexers/tvmaze/api.py
@@ -180,11 +180,29 @@ class TVmaze(BaseIndexer):
         """
         log.debug('Searching for show {0}', series)
 
-        results = self._show_search(series, request_language=self.config['language'])
-        if not results:
-            return
+        results = None
+        # If search term is digit's only, store it and attempt to search by id.
+        show_by_id = None
 
-        mapped_results = self._map_results(results, self.series_map, '|')
+        try:
+            if series.isdigit():
+                show_by_id = self._get_show_by_id(series, request_language=self.config['language'])
+            results = self._show_search(series, request_language=self.config['language'])
+        except IndexerShowNotFound:
+            pass
+
+        if not results and not show_by_id:
+            raise IndexerShowNotFound(
+                'Tvmaze show search failed in getting a result for search term {search}'.format(search=series)
+            )
+
+        mapped_results = []
+        if results:
+            mapped_results = self._map_results(results, self.series_map, '|')
+
+        # The search by id result, is already mapped. We can just add it to the array with results.
+        if show_by_id:
+            mapped_results.append(show_by_id['series'])
 
         return OrderedDict({'series': mapped_results})['series']