+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 #### New Features
+- Added Search shows by id ([#8058](https://github.com/pymedusa/Medusa/pull/8058))
 
 #### Improvements